// api/src/modules/user/constants/enums/errors.enum.ts
export enum ErrorMessage {
  USER_ID_EMPTY_ERROR = 'Bạn vui lòng chọn [người dùng].',
  ID_NAN_ERROR = 'Giá trị [id] không hợp lệ.',

  USER_NOT_FOUND_ERROR = '[Người dùng] không tồn tại (id: %s)',

  OPERATOR_USER_ERROR = 'Lưu thông tin người dùng thất bại.',

  NO_CONTENT = 'Không có dữ liệu hiển thị.',
  PERMISSION_NOT_FOUND_ERROR = 'Tính năng truy cập không tồn tại.',
}